Update url on scroll and nav click.
On nav click we push onto the stack so that the back button works as
expected. On scroll we only replace the top of the stack. So that the
back button takes you to the previous location that you clicked.
Clicking takes you to a new location and scrolling moves you around.